c48efdac83e642c6954e2b64707173fe40f29f54 fix(qt): avoid potential precision loss in amounts on Governance Tab (UdjinM6)

Pull request description:

  ## Issue being fixed or feature implemented
  https://github.com/dashpay/dash/pull/6403#discussion_r1847834174

  ## What was done?

  ## How Has This Been Tested?

  ## Breaking Changes

  ## Checklist:
  - [ ] I have performed a self-review of my own code
  - [ ] I have commented my code, particularly in hard-to-understand areas
  - [ ] I have added or updated relevant unit/integration/functional/e2e tests
  - [ ] I have made corresponding changes to the documentation
  - [ ] I have assigned this pull request to a milestone _(for repository code-owners and collaborators only)_

ACKs for top commit:
  knst:
    utACK c48efdac83e642c6954e2b64707173fe40f29f54
  PastaPastaPasta:
    utACK c48efdac83e642c6954e2b64707173fe40f29f54

Tree-SHA512: 705bc154e10150e32e3ff04d4a90de14e29ffd195cc9f94a753dd1b4588f2730d6526cd66d851005b956e0d074d39250d142baa946bea8df95dbc92718931762

+++ src/qt/governancelist.h
@@ -67,7 +67,7 @@ private:
     QString m_title;
     QDateTime m_startDate;
     QDateTime m_endDate;
-    float m_paymentAmount;
+    double m_paymentAmount;
     QString m_url;
 
 public:
@@ -76,7 +76,7 @@ public:
     QString hash() const;
     QDateTime startDate() const;
     QDateTime endDate() const;
-    float paymentAmount() const;
+    double paymentAmount() const;
     QString url() const;
     bool isActive() const;
     QString votingStatus(int nAbsVoteReq) const;
